backend_phase3/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
mcp==1.0.0
openai==1.3.5
sqlalchemy==2.0.23
httpx==0.25.2